[feature1] FEAT : BasicRepository 인터페이스 확장 및 구현으로 DB에서 CRUD하는 로직 담당하도록 클래스 작성
