+++ src/main/java/codesquad/springcafe/global/repository/BasicRepository.java
@@ -3,12 +3,12 @@ package codesquad.springcafe.global.repository;
 import java.util.Collection;
 import java.util.Optional;
 
-public interface Repository<T, ID> {
+public interface BasicRepository<T, ID> {
     Long save(T entity);
 
     Optional<T> findById(ID id);
 
     Collection<T> findAll();
 
-    Integer countAll();
+    Long countAll();
 }
